Stars feedback allowing 0.5 star reedback

+++ src/web/CCRS.WebApp.MVC/Models/FeedbackViewModel.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public int Stars { get; set; }
+        public float Stars { get; set; }
         public string Comment { get; set; }
         public string Image { get; set; }
         public DateTime CreatedAt { get; set; }
